Revert "xcm-emulator: move bridge tests under per-bridge crate"
This reverts commit 942092fa25069f8de7e2ccb378c640c2c17a694e.
