+++ Cargo.toml
@@ -67,7 +67,8 @@ members = [
 	"cumulus/parachains/common",
 	"cumulus/parachains/integration-tests/emulated/tests/assets/asset-hub-rococo",
 	"cumulus/parachains/integration-tests/emulated/tests/assets/asset-hub-westend",
-	"cumulus/parachains/integration-tests/emulated/tests/bridges/rococo-westend",
+	"cumulus/parachains/integration-tests/emulated/tests/bridges/bridge-hub-rococo",
+	"cumulus/parachains/integration-tests/emulated/tests/bridges/bridge-hub-westend",
 	"cumulus/parachains/integration-tests/emulated/common",
 	"cumulus/parachains/integration-tests/emulated/chains/relays/rococo",
 	"cumulus/parachains/integration-tests/emulated/chains/relays/westend",
