+++ cumulus/parachains/integration-tests/emulated/tests/bridges/bridge-hub-rococo/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "bridge-hub-rococo-integration-tests"
+version = "1.0.0"
+authors.workspace = true
+edition.workspace = true
+license = "Apache-2.0"
+description = "Bridge Hub Rococo runtime integration tests with xcm-emulator"
+publish = false
+
+[dependencies]
+codec = { package = "parity-scale-codec", version = "3.4.0", default-features = false }
+
+# Substrate
+frame-support = { path = "../../../../../../../substrate/frame/support", default-features = false}
+pallet-message-queue = { path = "../../../../../../../substrate/frame/message-queue" }
+
+# Polkadot
+xcm = { package = "staging-xcm", path = "../../../../../../../polkadot/xcm", default-features = false}
+pallet-xcm = { path = "../../../../../../../polkadot/xcm/pallet-xcm", default-features = false}
+xcm-executor = { package = "staging-xcm-executor",  path = "../../../../../../../polkadot/xcm/xcm-executor", default-features = false}
+
+# Bridges
+pallet-bridge-messages = { path = "../../../../../../../bridges/modules/messages", default-features = false}
+bp-messages = { path = "../../../../../../../bridges/primitives/messages", default-features = false}
+
+# Cumulus
+asset-test-utils = { path = "../../../../../../parachains/runtimes/assets/test-utils" }
+parachains-common = { path = "../../../../../../parachains/common" }
+cumulus-pallet-xcmp-queue = { path = "../../../../../../pallets/xcmp-queue", default-features = false}
+cumulus-pallet-dmp-queue = { path = "../../../../../../pallets/dmp-queue", default-features = false}
+bridge-hub-rococo-runtime = { path = "../../../../../../parachains/runtimes/bridge-hubs/bridge-hub-rococo", default-features = false }
+emulated-integration-tests-common = { path = "../../../common", default-features = false}
+rococo-westend-system-emulated-network = { path = "../../../networks/rococo-westend-system" }
