+++ cumulus/parachains/integration-tests/emulated/tests/bridges/bridge-hub-westend/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "rococo-westend-bridge-integration-tests"
+name = "bridge-hub-westend-integration-tests"
 version = "1.0.0"
 authors.workspace = true
 edition.workspace = true
 license = "Apache-2.0"
-description = "Rococo<>Westend bridge integration tests with xcm-emulator"
+description = "Bridge Hub Westend runtime integration tests with xcm-emulator"
 publish = false
 
 [dependencies]
@@ -28,7 +28,6 @@ asset-test-utils = { path = "../../../../../../parachains/runtimes/assets/test-u
 parachains-common = { path = "../../../../../../parachains/common" }
 cumulus-pallet-xcmp-queue = { path = "../../../../../../pallets/xcmp-queue", default-features = false}
 cumulus-pallet-dmp-queue = { path = "../../../../../../pallets/dmp-queue", default-features = false}
-bridge-hub-rococo-runtime = { path = "../../../../../../parachains/runtimes/bridge-hubs/bridge-hub-rococo", default-features = false }
 bridge-hub-westend-runtime = { path = "../../../../../../parachains/runtimes/bridge-hubs/bridge-hub-westend", default-features = false }
 emulated-integration-tests-common = { path = "../../../common", default-features = false}
 rococo-westend-system-emulated-network = { path = "../../../networks/rococo-westend-system" }
