+++ cumulus/parachains/integration-tests/emulated/tests/bridges/bridge-hub-westend/src/tests/teleport.rs
@@ -14,31 +14,17 @@
 // limitations under the License.
 
 use crate::*;
-use bridge_hub_rococo_runtime::xcm_config::XcmConfig as BridgeHubRococoXcmConfig;
-use bridge_hub_westend_runtime::xcm_config::XcmConfig as BridgeHubWestendXcmConfig;
+use bridge_hub_westend_runtime::xcm_config::XcmConfig;
 
 #[test]
-fn bridge_hub_rococo_teleport_to_other_system_parachains_works() {
+fn teleport_to_other_system_parachains_works() {
 	let amount = BRIDGE_HUB_ROCOCO_ED * 100;
 	let native_asset: MultiAssets = (Parent, amount).into();
 
 	test_parachain_is_trusted_teleporter!(
-		BridgeHubRococo,          // Origin
-		BridgeHubRococoXcmConfig, // XCM configuration
-		vec![AssetHubRococo],     // Destinations
-		(native_asset, amount)
-	);
-}
-
-#[test]
-fn bridge_hub_westend_teleport_to_other_system_parachains_works() {
-	let amount = BRIDGE_HUB_WESTEND_ED * 100;
-	let native_asset: MultiAssets = (Parent, amount).into();
-
-	test_parachain_is_trusted_teleporter!(
-		BridgeHubWestend,          // Origin
-		BridgeHubWestendXcmConfig, // XCM configuration
-		vec![AssetHubWestend],     // Destinations
+		BridgeHubWestend,      // Origin
+		XcmConfig,             // XCM configuration
+		vec![AssetHubWestend], // Destinations
 		(native_asset, amount)
 	);
 }
